Disable Next button if syncing is not enabled. (#1303)
This also:

* Makes us consistent by referring to the step as "configure-step" rather
  than a mix of "configure-step" and "configuration-step" in various parts
  of the code.
* Factors out the configure-step into its own module that can be tested
  separately, and adds some tests.
