+++ kitsune/sumo/static/sumo/scss/form-wizard-configure-step.styles.scss
